setup: server organization and server.js changes
