+++ server/package.json
@@ -2,12 +2,11 @@
   "name": "server",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
     "watch": "tsc -w",
-    "dev": "nodemon src/index.js",
-    "start": "node src/index.js",
-    "start2": "ts-node src/index.ts",
+    "dev": "nodemon src/server.js",
+    "start": "node src/server.js",
     "build": "tsc",
     "gen-env": "gen-env-types .env -o src/env.d.ts -e ."
   },
@@ -15,8 +14,6 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/express": "^4.17.17",
-    "@types/node-fetch": "^2.6.2",
     "gen-env-types": "^1.3.4",
     "nodemon": "^2.0.20",
     "ts-node": "^10.9.1"
@@ -26,7 +23,6 @@
     "cors": "^2.8.5",
     "dotenv-safe": "^8.2.0",
     "express": "^4.18.2",
-    "ioredis": "^5.3.1",
-    "tsconfig.json": "^1.0.11"
+    "ioredis": "^5.3.1"
   }
 }
