--- server/src/env.d.ts
@@ -1,9 +1,0 @@
-declare global {
-  namespace NodeJS {
-    interface ProcessEnv {
-      
-    }
-  }
-}
-
-export {}
